+++ CHANGELOG.md
@@ -8,6 +8,7 @@ with the current date and the next changes should go under a **[Next]** header.
 ## [Next]
 
 * Reorganize directory structure. ([@nwalters512](https://github.com/nwalters512) in [#108](https://github.com/illinois/queue/pull/108))
+* Remove location from notifications if the queue is a fixed-location queue. ([@redsn0w422](https://github.com/redsn0w422) in [#123](https://github.com/illinois/queue/pull/123))
 * Enter keypress adds new question, new queue, or new course. ([@emilojkovic](https://github.com/emilojkovic) in [#126](https://github.com/illinois/queue/pull/126))
 
 ## 19 April 2018